+++ CHANGELOG.md
@@ -14,10 +14,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 * Circular symlinks no longer cause infinite loops when syncing a folder
+* Fix crash on upload retry with unbound data source
 
 ### Infrastructure
 * Replaced `pyflakes` with `ruff` for linting
 * Refactored logic for resuming large file uploads to unify code paths, correct inconsistencies, and enhance configurability (#381)
+* Automatically set copyright date when generating the docs
 
 ## [1.21.0] - 2023-04-17